backend/requirements.txt
openai-whisper==20231117
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
torch==2.0.1
torchaudio==2.0.2
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10